+++ p2128/print.hpp
@@ -10,7 +10,7 @@ template<typename T>
 void print_vec(std::vector<T> v) {
     std::cout<<'[';
     if (v.size() > 0) {
-        for (int i = 0; i < v.size()-1; i++) {
+        for (size_t i = 0; i < v.size()-1; i++) {
             std::cout<<v[i]<<", ";
         }
         std::cout<<v[v.size()-1];
@@ -22,7 +22,7 @@ template<typename T>
 void print_vec(std::vector<std::vector<T>> v) {
     std::cout<<'[';
     if (v.size() >  0) {
-        for (int i = 0; i < v.size()-1; i++) {
+        for (size_t i = 0; i < v.size()-1; i++) {
             print_vec(v[i]);
             std::cout<<", ";
         }
