+++ p5460/print.hpp
@@ -31,7 +31,7 @@ template<typename T>
 void print_vec(const std::vector<tree_ptr<T>>& v) {
     std::cout<<'[';
     if (v.size() > 0) {
-        for (int i = 0; i < v.size()-1; i++) {
+        for (size_t i = 0; i < v.size()-1; i++) {
             print_tree(v[i].get());
             std::cout << ',' << std::endl;
         }
@@ -44,7 +44,7 @@ template<typename T, typename U>
 void print_vec(const std::vector<std::pair<T, U>>& v) {
     std::cout<<'[';
     if (v.size() > 0) {
-        for (int i = 0; i < v.size()-1; i++) {
+        for (size_t i = 0; i < v.size()-1; i++) {
             std::cout<<'('<<v[i].first<<','<<v[i].second<<')'<<", ";
         }
         std::cout<<'('<<v[v.size()-1].first<<','<<v[v.size()-1].second<<')';
@@ -56,7 +56,7 @@ template<typename T>
 void print_vec(const std::vector<T>& v) {
     std::cout<<'[';
     if (v.size() > 0) {
-        for (int i = 0; i < v.size()-1; i++) {
+        for (size_t i = 0; i < v.size()-1; i++) {
             std::cout<<v[i]<<", ";
         }
         std::cout<<v[v.size()-1];
@@ -68,7 +68,7 @@ template<typename T>
 void print_vec(const std::vector<std::vector<T>>& v) {
     std::cout<<'[';
     if (v.size() >  0) {
-        for (int i = 0; i < v.size()-1; i++) {
+        for (size_t i = 0; i < v.size()-1; i++) {
             print_vec(v[i]);
             std::cout<<", ";
         }
